net: lib: lwm2m_client_utils: Move location assistance events
Move location assistance events to separate folder. This way
applications that doesn't use the event listener for location
objects don't have to include the event definitions.

This is convenient when creating that will not need to include
mocking for the event manager.
