+++ include/net/lwm2m_client_utils_location.h
@@ -18,8 +18,6 @@
 #include <zephyr/kernel.h>
 #include <nrf_modem_gnss.h>
 #include <zephyr/net/lwm2m.h>
-#include <app_event_manager.h>
-#include <app_event_manager_profiler_tracer.h>
 
 #define LOCATION_ASSIST_NEEDS_UTC			BIT(0)
 #define LOCATION_ASSIST_NEEDS_EPHEMERIES		BIT(1)
@@ -73,38 +71,6 @@ int location_assistance_pgps_request_send(struct lwm2m_ctx *ctx, bool confirmabl
 #endif
 
 #if defined(CONFIG_LWM2M_CLIENT_UTILS_LOCATION_ASSIST_EVENTS)
-#if defined(CONFIG_LWM2M_CLIENT_UTILS_LOCATION_ASSIST_AGPS)
-struct gnss_agps_request_event {
-	struct app_event_header header;
-
-	struct nrf_modem_gnss_agps_data_frame agps_req;
-};
-
-APP_EVENT_TYPE_DECLARE(gnss_agps_request_event);
-
-#endif /* CONFIG_LWM2M_CLIENT_UTILS_LOCATION_ASSIST_AGPS */
-#if defined(CONFIG_LWM2M_CLIENT_UTILS_LOCATION_ASSIST_CELL)
-struct cell_location_request_event {
-	struct app_event_header header;
-};
-
-APP_EVENT_TYPE_DECLARE(cell_location_request_event);
-
-struct cell_location_inform_event {
-	struct app_event_header header;
-};
-
-APP_EVENT_TYPE_DECLARE(cell_location_inform_event);
-#endif /* CONFIG_LWM2M_CLIENT_UTILS_LOCATION_ASSIST_CELL */
-
-#if defined(CONFIG_LWM2M_CLIENT_UTILS_LOCATION_ASSIST_PGPS)
-struct pgps_data_request_event {
-	struct app_event_header header;
-};
-
-APP_EVENT_TYPE_DECLARE(pgps_data_request_event);
-#endif /* CONFIG_LWM2M_CLIEN_UTILS_LOCATION_ASSIST_PGPS */
-
 int location_event_handler_init(struct lwm2m_ctx *ctx);
 #endif /* CONFIG_LWM2M_CLIENT_UTILS_LOCATION_ASSIST_EVENTS */
