+++ subsys/net/lib/lwm2m_client_utils/location/location_events.c
@@ -5,6 +5,7 @@
  */
 
 #include <net/lwm2m_client_utils_location.h>
+#include <net/lwm2m_client_utils_location_events.h>
 
 #include <stdio.h>
